pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "selenium-mcp"
version = "1.0.0"
description = "MCP server integrating Selenium 4 (BiDi) for browser automation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "selenium>=4.20.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "webdriver-manager>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
selenium-mcp = "server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "driver*", "events*", "models*", "tools*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4", "PT"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
